[REFACTOR(ReservationPresenter)] 안드로이드 의존성 제거
